+++ apps/api/src/interfaces/http/containers/containers-http.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { ContainersUseCasesModule } from '../../../application/use-cases/containers/containers-use-cases.module';
+import { ContainersController } from './containers.controller';
+
+@Module({
+  imports: [ContainersUseCasesModule],
+  controllers: [ContainersController],
+})
+export class ContainersHttpModule {}